Update README.md
Problem Vision For Blind Aims for:
According to the World Health Organization (WHO), around 285 million people around the world are estimated to be visually impaired. Due to blindness visually impaired people are not able to recognize the people around them and is the major issue for blind population which are facing many problems, also people with disabilities are not able to control home appliance by their own which too leads to an another problem in itself.

My Solutiom: Vision for Blind
So my project deals with the aim of solving disability to see for Blind People as it recognises objects and faces of the people around them by using Image processing and Computer Vision algorithm which is nothing but Facial Recognition without using pre-defined Facial Recognition library of Python.As of my project first create the database of a person then whenever that person comes in the range of camera it recognise them. After recognition takes place where the final information is received which is then converted into speech or audio output by using different speech engine such as gTTS, E-Speak though gTTs is more accurate than TTs.( I used pyytts)